% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ash_on_maskedZ.R
\name{ash_on_maskedZ}
\alias{ash_on_maskedZ}
\title{run ash condition by condition on masked Z scores}
\usage{
ash_on_maskedZ(
  Z,
  P = NULL,
  p.thresh = 0.5,
  Shat = NULL,
  verbose = FALSE,
  eps = 1e-10
)
}
\description{
The procedure masks the z scores, by taking the bigger one(absolute value) in masked set, then fit ash condition by condition
}
