% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_obj_and_posterior.R
\name{calc_obj_maskedmd}
\alias{calc_obj_maskedmd}
\title{evaluate objective function}
\usage{
calc_obj_maskedmd(lik.list, Z.comb, Pi, N, R, multiplier = NULL)
}
\description{
evaluate objective function
}
