% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_obj_and_posterior.R
\name{calc_post_summary}
\alias{calc_post_summary}
\title{calculate posterior summaries, given data and prior parameters.}
\usage{
calc_post_summary(Z.comb, Ulist, Pi, post_weights, pi_thresh)
}
\description{
calculate posterior summaries, given data and prior parameters.
}
