% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_obj_and_posterior.R
\name{calc_post_weights}
\alias{calc_post_weights}
\title{calculate posterior weights}
\usage{
calc_post_weights(Z.comb, lik.list, Pi)
}
\arguments{
\item{Z.comb}{augmented z score matrices}

\item{lik.list}{output of `calc_relative_lik_matrix` from mashr}

\item{Pi}{prior weights}
}
\value{
posterior weights
}
\description{
calculate posterior weights
}
