% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{enumerate.z}
\alias{enumerate.z}
\title{returns all possible combinations of masked z-scores and |det(Jacobian)|}
\usage{
enumerate.z(z, thresh)
}
\arguments{
\item{z}{a vector}

\item{thresh}{|z| larger than thresh or |z|<=s(t) will be masked. when thresh<=Phi^{-1}(0.75), all masked.}
}
\value{
a J*R matrix of all possible z scores, a length J vector of |det(Jacobian)|, and masked z scores.
}
\description{
returns all possible combinations of masked z-scores and |det(Jacobian)|
}
