% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maskedmash.R
\name{estimate_pi}
\alias{estimate_pi}
\title{Estimate prior weights using masked data, given covariance matrices.}
\usage{
estimate_pi(
  Z.comb,
  U.canon,
  U.data,
  grid,
  usepointmass,
  Pi,
  max_iter,
  tol,
  pi_thresh,
  normalizeU,
  algorithm.version,
  optmethod,
  verbose,
  printevery,
  control,
  prior
)
}
\description{
Estimate prior weights using masked data, given covariance matrices.
}
