% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maskedmashFDR.R
\name{fdp.hat}
\alias{fdp.hat}
\title{Estimate of FDP as in adapt paper, without threshold version}
\usage{
fdp.hat(rej.set, P)
}
\description{
Estimate of FDP as in adapt paper, without threshold version
}
