% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fdp}
\alias{fdp}
\title{False Discovery Proportion}
\usage{
fdp(rej.idx, true.signal.idx)
}
\arguments{
\item{rej.idx}{index of rejections}

\item{true.signal.idx}{index of true signals}
}
\description{
False Discovery Proportion
}
