% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fsp}
\alias{fsp}
\title{False Sign Proportion}
\usage{
fsp(rej.sign, true.sign)
}
\arguments{
\item{rej.sign}{sign of findings}

\item{true.sign}{true sign of signals, whose index corresponds to those rejected ones.}
}
\description{
False Sign Proportion
}
