% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mash_on_maskedZ.R
\name{mash_on_maskedZ}
\alias{mash_on_maskedZ}
\title{mash on masked Z scores}
\usage{
mash_on_maskedZ(
  Z,
  P = NULL,
  p.thresh = 0.5,
  npc = 5,
  Shat = NULL,
  verbose = FALSE,
  eps = 1e-10
)
}
\description{
The procedure masks the z scores, by taking the bigger one(absolute value) in masked set, then fit mash.
}
