% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mash_wrapper.R
\name{mash_wrapper}
\alias{mash_wrapper}
\title{Run mash}
\usage{
mash_wrapper(
  Bhat,
  P = NULL,
  npc = 5,
  Shat = NULL,
  nu = ncol(Bhat) + 1,
  verbose = FALSE,
  adjust = "lb",
  md.method = "teem",
  strong.thresh = 0.2,
  U.true = NULL,
  U.data = NULL,
  n.data = NULL,
  adj.const = 0.001
)
}
\arguments{
\item{adjust}{how to adjust the covariance matrix, either prior or lb or const. lb is for lower bound}
}
\description{
Run mash
}
