% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mashFDR.R
\name{mashFDR}
\alias{mashFDR}
\title{mash for FDR control}
\usage{
mashFDR(obj, alpha = 0.05)
}
\arguments{
\item{obj}{fitted mash object}

\item{alpha}{target FDR level}
}
\description{
order the tests from most to least significant based on lfsr, and reject least significant ones until FDP<alpha
}
