% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maskedmash.R
\name{masked.mash}
\alias{masked.mash}
\title{masked mash}
\usage{
masked.mash(
  data,
  thresh = NULL,
  Pi = NULL,
  U.canon = NULL,
  U.data,
  fixg = FALSE,
  usepointmass = TRUE,
  pi_thresh = 1e-05,
  gridmult = sqrt(2),
  grid = NULL,
  normalizeU = TRUE,
  algorithm.version = "Rcpp",
  optmethod = "mixSQP",
  prior = "nullbiased",
  max_iter = 1000,
  tol = 1e-05,
  verbose = TRUE,
  printevery = 100,
  control = list()
)
}
\arguments{
\item{data}{object from mash_set_data}

\item{thresh}{NULL or a number, |z-score| larger than thresh will be masked.}

\item{Pi}{prior weights; either fixed or as init value}

\item{U.canon}{a list of canonical prior cov matrices; fixed.}

\item{U.data}{a list of data-driven prior cov marrices; either fixed or as initialization.}

\item{fixg}{whether fix the prior weights and covariance matrices.}

\item{usepointmass}{whether to include a point mass at 0}

\item{pi_thresh}{threshold below which mixture components are ignored}

\item{normalizeU}{whether normalize U such that the largest diag element is 1}

\item{algorithm.version}{Rcpp or R for evaluating likelihood}

\item{optmethod}{'mixSQP' or 'EM'}

\item{prior}{nullbiased or uniform}

\item{verbose}{TRUE to print progress}

\item{control}{a list of control parameters for SQP}
}
\value{
a list of Posterior mean, sd, lfsr, lfdr, negativeProb.
}
\description{
masked mash
}
