% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maskedmash.R
\name{maskedmash_wrapper}
\alias{maskedmash_wrapper}
\title{Wrapper function for running masked mash}
\usage{
maskedmash_wrapper(
  Z,
  P = NULL,
  p.thresh = 0.5,
  npc = 5,
  adjust = "lb",
  adj.const = 0.001,
  verbose = FALSE,
  strong_for_md = TRUE,
  strong.thresh = 0.2,
  usepointmass = TRUE,
  return_post_weights = FALSE
)
}
\arguments{
\item{Z}{matrix of Z scores}

\item{P}{matrix of p values, can be NULL, and will be calculated from Z scores}

\item{p.thresh}{threshold of p-values for masking. It take values in [0,0.5]. p<=thresh and p>=(1-thresh) will be masked.}

\item{npc}{number of data driven covariance matrices for mash}

\item{adjust}{method for adjusting diagonal of estimated covariance matrices, "lb" or "prior".}

\item{strong_for_md}{Whether use strong effects for estimating data driven cov matrices}

\item{strong.thresh}{for deciding strong effects.}
}
\description{
Wrapper function for running masked mash
}
