% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{powr}
\alias{powr}
\title{Power of statistical procedure}
\usage{
powr(rej.idx, true.signal.idx)
}
\arguments{
\item{rej.idx}{index of rejections}

\item{true.signal.idx}{index of true signals}
}
\description{
Power of statistical procedure
}
