% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simDataI.ult}
\alias{simDataI.ult}
\title{simulate Z scores}
\usage{
simDataI.ult(
  N,
  Ulist,
  Pi = NULL,
  prior = "t",
  t.df = 10,
  half.uniform = FALSE,
  mean.range = 4
)
}
\arguments{
\item{N}{total number of samples}

\item{Ulist}{a list of correlation matrices}

\item{Pi}{prior weights}

\item{prior}{normal, t, uniform}

\item{half.uniform}{True to generate only positive uniforms}

\item{mean.range}{range of alternative E(z), from -mean.range to mean.range}
}
\description{
simulate Z scores
}
